projects/pontos_turisticos/requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.6
django-filter==2.2.0
django-reset-migrations==0.3.1
djangorestframework==3.10.3
Markdown==3.1.1
Pillow==6.2.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.3
sqlparse==0.3.0
static3==0.7.0
